Uploading the IMR source code as requested by Ryan.
This will allow editing of the script for UTF-8 usage.

dbconnect.php is a blank shell of a script, and is there merely for the purpose of showing the structure, the actual SQL connection details are in the one on the HoverRace.com website (naturally).

My apologies, it's not the prettiest :).